now terminating virtual display process if an exception occurs (had lotos of hanging Xvfb processes)

+++ border_wait_times.py
@@ -44,7 +44,8 @@ def scrape(url, elem_ids):
         if 'darwin' not in sys.platform:
             display.stop()
     except Exception as e:
-        display.popen.terminate()
+        if 'darwin' not in sys.platform:
+            display.popen.terminate()
         return None
     return src
